requirements.txt
# Core Dependencies
streamlit>=1.28.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
spacy>=3.7.0
nltk>=3.8.0
pyyaml>=6.0.0
jinja2>=3.1.0

# AI/ML Dependencies
tensorflow>=2.14.0
efficientnet>=1.1.1
scikit-learn>=1.3.0
scipy>=1.11.0
realesrgan>=0.3.0
basicsr>=1.4.2

# Optional Enhancement
matplotlib>=3.8.0
pandas>=2.1.0

# Utilities
python-dateutil>=2.8.0
